Divide RenderableLayers
