+++ packages/flame_tiled/lib/src/renderable_layers/object_layer.dart
@@ -6,12 +6,12 @@ import 'package:tiled/tiled.dart';
 
 @internal
 class ObjectLayer extends RenderableLayer<ObjectGroup> {
-  ObjectLayer(
-    super.layer,
-    super.parent,
-    super.map,
-    super.destTileSize,
-  );
+  ObjectLayer({
+    required super.layer,
+    required super.parent,
+    required super.map,
+    required super.destTileSize,
+  });
 
   @override
   void render(Canvas canvas, Camera? camera) {
@@ -27,7 +27,12 @@ class ObjectLayer extends RenderableLayer<ObjectGroup> {
     TiledMap map,
     Vector2 destTileSize,
   ) async {
-    return ObjectLayer(layer, null, map, destTileSize);
+    return ObjectLayer(
+      layer: layer,
+      parent: null,
+      map: map,
+      destTileSize: destTileSize,
+    );
   }
 
   @override
